// XuReverseProxy.Frontend/src/generated/Models/Web/LoginRequest.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface LoginRequest
{
	username: string;
	password: string;
	totp: string;
	recoveryCode: string;
	returnPath: string;
}